+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Send `paypal_context_id` in `batch_params` to PayPal's analytics service (FPTI) when available
 * BraintreeVenmo
   * Add `isFinalAmount` to `BTVenmoRequest`
+  * Add `BTVenmoRequest.fallbackToWeb`
+    * If set to `true` customers will fallback to a web based Venmo flow if the Venmo app is not installed
+    * This method uses Universal Links instead of URL Schemes
 
 ## 6.12.0 (2024-01-18)
 * BraintreePayPal